Move websocket proxy tests to existing test file
Consolidate all websocket tests in tests/_internal/test_websockets.py
rather than creating a new file. This keeps related tests together
and maintains consistency with the existing test structure.
